+++ packages/flutter_wright_sdk/test/mock_provider_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_visual_loop/flutter_visual_loop.dart';
+import 'package:flutter_wright_sdk/flutter_wright_sdk.dart';
 
 void main() {
   group('InMemoryMockDataProvider', () {